requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
openai>=1.35.0
httpx>=0.27.0
python-multipart>=0.0.9
requests>=2.31.0
google-genai>=1.15.0